Create new standalone employee form to use inside of employee create component
