+++ src/components/EmployeeDetails.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
-import { Picker, Text } from 'react-native';
-import { Card, CardSection, Input, Button } from './common';
+import { Card, CardSection, Button } from './common';
 import { connect } from 'react-redux';
 import { employeeUpdate, employeeDetails } from '../actions';
+import EmployeeForm from './EmployeeForm';
 
 class EmployeeDetails extends Component {
   onButtonPress() {
@@ -18,46 +18,7 @@ class EmployeeDetails extends Component {
   render() {
     return (
       <Card>
-        <CardSection>
-          <Input
-            label="Name"
-            placeholder="Name"
-            value={this.props.name}
-            onChangeText={value =>
-              this.props.employeeUpdate({ prop: 'name', value })
-            }
-          />
-        </CardSection>
-
-        <CardSection>
-          <Input
-            label="Phone"
-            placeholder="555-111-1234"
-            value={this.props.phone}
-            onChangeText={value =>
-              this.props.employeeUpdate({ prop: 'phone', value })
-            }
-          />
-        </CardSection>
-
-        <CardSection style={{ flexDirection: 'column' }}>
-          <Text style={styles.pickerLabelStyle}>Select a Shift</Text>
-          <Picker
-            selectedValue={this.props.shift}
-            onValueChange={value =>
-              this.props.employeeUpdate({ prop: 'shift', value })
-            }
-          >
-            <Picker.Item label="Monday" value="Monday" />
-            <Picker.Item label="Tuesday" value="Tuesday" />
-            <Picker.Item label="Wednesday" value="Wednesday" />
-            <Picker.Item label="Thursday" value="Thursday" />
-            <Picker.Item label="Friday" value="Friday" />
-            <Picker.Item label="Saturday" value="Saturday" />
-            <Picker.Item label="Sunday" value="Sunday" />
-          </Picker>
-        </CardSection>
-
+        <EmployeeForm {...this.props} />
         <CardSection>
           <Button buttonText="Create" onPress={this.onButtonPress.bind(this)} />
         </CardSection>
@@ -65,13 +26,6 @@ class EmployeeDetails extends Component {
     );
   }
 }
-
-const styles = {
-  pickerLabelStyle: {
-    fontSize: 18,
-    paddingLeft: 20
-  }
-};
 
 const mapStateToProps = state => {
   const { name, phone, shift } = state.employeeForm;
