+++ src/components/ListItem.js
@@ -1,15 +1,24 @@
 import React, { Component } from 'react';
-import { Text } from 'react-native';
+import { Text, TouchableWithoutFeedback, View } from 'react-native';
 import { CardSection } from './common';
+import { Actions } from 'react-native-router-flux';
 
 class ListItem extends Component {
+  onRowPress() {
+    Actions.employeeDetails({ employee: this.props.employee });
+  }
+
   render() {
     const { name } = this.props.employee;
 
     return (
-      <CardSection>
-        <Text style={styles.titleStyle}>{name}</Text>
-      </CardSection>
+      <TouchableWithoutFeedback onPress={this.onRowPress.bind(this)}>
+        <View>
+          <CardSection>
+            <Text style={styles.titleStyle}>{name}</Text>
+          </CardSection>
+        </View>
+      </TouchableWithoutFeedback>
     );
   }
 }
